+++ src/main/kotlin/cz/machovec/lekovyportal/core/dto/distribution/AggregateDistributorProductMovementCountDto.kt
@@ -2,10 +2,11 @@ package cz.machovec.lekovyportal.core.dto.distribution
 
 import cz.machovec.lekovyportal.core.domain.distribution.DistributorPurchaserType
 import cz.machovec.lekovyportal.core.domain.distribution.MovementType
+import java.math.BigDecimal
 
 data class AggregateDistributorProductMovementCountDto(
     val medicinalProductId: Long,
     val purchaserType: DistributorPurchaserType,
     val movementType: MovementType,
-    val packageCount: Long
+    val packageCount: BigDecimal
 )
